pyproject.toml
[project]
name = "stac-planet-api"
version = "0.1.1"
description = ""
authors = [
    { name = "Rhys Evans", email = "rhys.r.evans@stfc.ac.uk" },
    { name = "Guyon Duifhuizen", email = "guyon.duifhuizen@planet.com" },
]
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.13,<3.14"
classifiers = [
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "cryptography>=43.0.0",
    "fastapi>=0.111.1",
    "httpx>=0.27.0",
    "orjson>=3.10.6",
    "pygeofilter>=0.2.4",
    "pydantic-settings>=2.4.0",
    "stac-fastapi-api==3.0.0b2",
    "stac-fastapi-extensions==3.0.0b2",
    "stac-fastapi-types==3.0.0b2",
    "stac-pydantic>=3.1.1",
    "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pyright",
    "pytest",
    "pytest-mock",
    "pytest-watcher",
    "pytest-xdist",
    "ruff",
    "validate-pyproject",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
    "E",    # Pycodestyle errors
    "W",    # Pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "FAST", # FastAPI
    "RUF",  # RUF-specific rules
    "SIM",  # Simplifications
    "B",    # Bugbear
    "ANN",  # Annotations
    "PLC",  # Pylint convention
    "PLE",  # Pylint errors
    "UP",   # pyupgrade
    "FURB", # refurb
    "PT",   # pytest
]
ignore = [
    "E501",   # line too long
    "SIM108", # ternary operator
    "SIM103", # return bool directly
    "B008",   # function call in default argument (incompatible with FastAPI)
]
